Share variable names for easier reading/writing
With the previous version, some variables names were left out of
the CSV header because of the line length. Specifying a longer
value for LRECL (logical record length) should fix that.
Unfortunately PROC IMPORT has a hard-coded maximum line length
which is too short for our names, so it cannot be used to read
the data directly.
Using a %LET macro to hold the names of variables lets both files
refer to them without needing to copy the names in both places.