trying to use the variables from the dataset
more demo code. So we want to pass on variables from the model_runner to the model_consumer and then use the same variables in the consumer.
On reflection: the call symput statement in the runner doesn't matter. I will need to save the data in a dataset anyway it seems. Or does Symput store the data locally? Like system parameters?