Don't invalidate newp in first time period
The reworking of the sex worker code assumes that newp_tm1 is not
missing. By setting newp to . in the first period, newp_tm1 will be
. in the second period, causing an error for sex workers. This
should not affect the program, as newp_tm1 is only used for t>=2.

We could do the same for the other _tm1 variables here, but I
couldn't verify that they are not used for t=1.