Discounting
Modified the formula for discounting on both files. I removed the variable discount_adj (it was set to 1). The program tends to remove "s_" when a variable is multiplied  by the scale factor and by 4, there are a few occasions where this is not the case and I wonder whether it should for consistency. I have replaced sf_2019 with a macro variable &sf. Lines 164-182 sf_2020 was used instead of sf_2019. I have not tried running create_wide_file but happy to do so if you suggest which output file I could use.