Update hiv_synthesis.sas
removing the < 65 condition as discussed - note some of the s_ will now include people aged 65+ when that was not the case previously.  I think I'm ready for us to make this change and check the Table 1.  But I'm happy to wait before I merge this until you are all comfortable.