update
I think with these changes we ensure that the model simulation realization for any given seed does not change when we change the sum statement to age ge 15 from age 15-64.  The variables fed into the next period relate to transmission and should be based on the age 15-64 population so I have modified the definition of some variables and added some new ones.