Update hiv_synthesis.sas
tidying up some outdated code.  There is an independent effect of being on ART for risk of clinical disease.   Previously we had separate effects for people on 1 or 2 drug regimens but that related to when these regimens were still used but are known to be inadequate.  So no longer relevant.  I also had created a parameter name for death but not WHO stage 3 and 4.  Lastly, adding some code for re_option_104.